refactor: modularize by function
