+++ packages/eip-5792-middleware/jest.config.js
@@ -17,7 +17,7 @@ module.exports = merge(baseConfig, {
   // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
     global: {
-      branches: 90,
+      branches: 85,
       functions: 100,
       lines: 90,
       statements: 90,
