+++ packages/eip-5792-middleware/src/index.ts
@@ -1,8 +1,5 @@
-export {
-  processSendCalls,
-  getCallsStatus,
-  getCapabilities,
-  AtomicCapabilityStatus,
-} from './eip5792';
-
-export type { EIP5792Messenger } from './eip5792';
+export { processSendCalls } from './processSendCalls';
+export { getCallsStatus } from './getCallsStatus';
+export { getCapabilities } from './getCapabilities';
+export { AtomicCapabilityStatus } from './constants';
+export type { EIP5792Messenger } from './types';
