+++ packages/eip-5792-middleware/src/processSendCalls.ts
@@ -1,38 +1,17 @@
 import type {
-  AccountsControllerGetSelectedAccountAction,
-  AccountsControllerGetStateAction,
-} from '@metamask/accounts-controller';
-import type { Messenger } from '@metamask/base-controller';
-import type {
-  GetCallsStatusResult,
-  GetCapabilitiesResult,
   SendCalls,
   SendCallsResult,
 } from '@metamask/eth-json-rpc-middleware';
-import { GetCallsStatusCode } from '@metamask/eth-json-rpc-middleware';
 import type { KeyringTypes } from '@metamask/keyring-controller';
-import type {
-  NetworkControllerGetNetworkClientByIdAction,
-  NetworkControllerGetStateAction,
-} from '@metamask/network-controller';
-import type { PreferencesControllerGetStateAction } from '@metamask/preferences-controller';
 import { JsonRpcError, rpcErrors } from '@metamask/rpc-errors';
 import type {
   BatchTransactionParams,
-  IsAtomicBatchSupportedResult,
   IsAtomicBatchSupportedResultEntry,
-  Log,
   SecurityAlertResponse,
   TransactionController,
-  TransactionControllerGetStateAction,
-  TransactionMeta,
-  TransactionReceipt,
   ValidateSecurityRequest,
 } from '@metamask/transaction-controller';
-import {
-  TransactionEnvelopeType,
-  TransactionStatus,
-} from '@metamask/transaction-controller';
+import { TransactionEnvelopeType } from '@metamask/transaction-controller';
 import type { Hex, JsonRpcRequest } from '@metamask/utils';
 import { bytesToHex } from '@metamask/utils';
 import { parse, v4 as uuid } from 'uuid';
@@ -40,24 +19,11 @@ import { parse, v4 as uuid } from 'uuid';
 import {
   EIP5792ErrorCode,
   KEYRING_TYPES_SUPPORTING_7702,
-  MESSAGE_TYPE,
+  MessageType,
+  VERSION,
 } from './constants';
-
-type Actions =
-  | AccountsControllerGetStateAction
-  | AccountsControllerGetSelectedAccountAction
-  | NetworkControllerGetNetworkClientByIdAction
-  | TransactionControllerGetStateAction
-  | PreferencesControllerGetStateAction
-  | NetworkControllerGetStateAction;
-
-export type EIP5792Messenger = Messenger<Actions, never>;
-
-export enum AtomicCapabilityStatus {
-  Supported = 'supported',
-  Ready = 'ready',
-  Unsupported = 'unsupported',
-}
+import type { EIP5792Messenger } from './types';
+import { getAccountKeyringType } from './utils';
 
 type ProcessSendCallsHooks = {
   addTransactionBatch: TransactionController['addTransactionBatch'];
@@ -75,18 +41,6 @@ type ProcessSendCallsRequest = JsonRpcRequest & {
   networkClientId: string;
   origin?: string;
 };
-
-type GetCapabilitiesHooks = {
-  getDismissSmartAccountSuggestionEnabled: () => boolean;
-  getIsSmartTransaction: (chainId: Hex) => boolean;
-  isAtomicBatchSupported: TransactionController['isAtomicBatchSupported'];
-  isRelaySupported: (chainId: Hex) => Promise<boolean>;
-  getSendBundleSupportedChains: (
-    chainIds: Hex[],
-  ) => Promise<Record<string, boolean>>;
-};
-
-const VERSION = '2.0.0';
 
 /**
  * Processes a sendCalls request for EIP-5792 atomic transactions.
@@ -166,159 +120,6 @@ export async function processSendCalls(
 }
 
 /**
- * Retrieves the status of a transaction batch by its ID.
- *
- * @param messenger - Messenger instance for controller communication.
- * @param id - The batch ID to look up (hexadecimal string).
- * @returns GetCallsStatusResult containing the batch status, receipts, and metadata.
- * @throws JsonRpcError with EIP5792ErrorCode.UnknownBundleId if no matching bundle is found.
- */
-export function getCallsStatus(
-  messenger: EIP5792Messenger,
-  id: Hex,
-): GetCallsStatusResult {
-  const transactions = messenger
-    .call('TransactionController:getState')
-    .transactions.filter((tx) => tx.batchId === id);
-
-  if (!transactions?.length) {
-    throw new JsonRpcError(
-      EIP5792ErrorCode.UnknownBundleId,
-      `No matching bundle found`,
-    );
-  }
-
-  const transaction = transactions[0];
-  const { chainId, txReceipt: rawTxReceipt } = transaction;
-  const status = getStatusCode(transaction);
-  const txReceipt = rawTxReceipt as Required<TransactionReceipt> | undefined;
-  const logs = (txReceipt?.logs ?? []) as Required<Log>[];
-
-  const receipts: GetCallsStatusResult['receipts'] = txReceipt && [
-    {
-      blockHash: txReceipt.blockHash as Hex,
-      blockNumber: txReceipt.blockNumber as Hex,
-      gasUsed: txReceipt.gasUsed as Hex,
-      logs: logs.map((log: Required<Log> & { data: Hex }) => ({
-        address: log.address as Hex,
-        data: log.data,
-        topics: log.topics as unknown as Hex[],
-      })),
-      status: txReceipt.status as '0x0' | '0x1',
-      transactionHash: txReceipt.transactionHash,
-    },
-  ];
-
-  return {
-    version: VERSION,
-    id,
-    chainId,
-    atomic: true, // Always atomic as we currently only support EIP-7702 batches
-    status,
-    receipts,
-  };
-}
-
-/**
- * Retrieves the capabilities for atomic transactions on specified chains.
- *
- * @param hooks - Object containing required controller hooks and utilities.
- * @param hooks.getDismissSmartAccountSuggestionEnabled - Function to check if smart account suggestions are disabled.
- * @param hooks.getIsSmartTransaction - Function to check if a chain supports smart transactions.
- * @param hooks.isAtomicBatchSupported - Function to check if atomic batching is supported.
- * @param hooks.isRelaySupported - Function to check if relay is supported on a chain.
- * @param hooks.getSendBundleSupportedChains - Function to get chains that support send bundle.
- * @param messenger - Messenger instance for controller communication.
- * @param address - The account address to check capabilities for.
- * @param chainIds - Array of chain IDs to check capabilities for (if undefined, checks all configured networks).
- * @returns Promise resolving to GetCapabilitiesResult mapping chain IDs to their capabilities.
- */
-export async function getCapabilities(
-  hooks: GetCapabilitiesHooks,
-  messenger: EIP5792Messenger,
-  address: Hex,
-  chainIds: Hex[] | undefined,
-) {
-  const {
-    getDismissSmartAccountSuggestionEnabled,
-    getIsSmartTransaction,
-    isAtomicBatchSupported,
-    isRelaySupported,
-    getSendBundleSupportedChains,
-  } = hooks;
-
-  let chainIdsNormalized = chainIds?.map(
-    (chainId) => chainId.toLowerCase() as Hex,
-  );
-
-  if (!chainIdsNormalized?.length) {
-    const networkConfigurations = messenger.call(
-      'NetworkController:getState',
-    ).networkConfigurationsByChainId;
-    chainIdsNormalized = Object.keys(networkConfigurations) as Hex[];
-  }
-
-  const batchSupport = await isAtomicBatchSupported({
-    address,
-    chainIds: chainIdsNormalized,
-  });
-
-  const alternateGasFeesAcc = await getAlternateGasFeesCapability(
-    chainIdsNormalized,
-    batchSupport,
-    getIsSmartTransaction,
-    isRelaySupported,
-    getSendBundleSupportedChains,
-    messenger,
-  );
-
-  return chainIdsNormalized.reduce<GetCapabilitiesResult>((acc, chainId) => {
-    const chainBatchSupport = (batchSupport.find(
-      ({ chainId: batchChainId }) => batchChainId === chainId,
-    ) ?? {}) as IsAtomicBatchSupportedResultEntry & {
-      isRelaySupported: boolean;
-    };
-
-    const { delegationAddress, isSupported, upgradeContractAddress } =
-      chainBatchSupport;
-
-    const isUpgradeDisabled = getDismissSmartAccountSuggestionEnabled();
-    let isSupportedAccount = false;
-
-    try {
-      const keyringType = getAccountKeyringType(address, messenger);
-      isSupportedAccount = KEYRING_TYPES_SUPPORTING_7702.includes(keyringType);
-    } catch (error) {
-      // Intentionally empty
-    }
-
-    const canUpgrade =
-      !isUpgradeDisabled &&
-      upgradeContractAddress &&
-      !delegationAddress &&
-      isSupportedAccount;
-
-    if (!isSupported && !canUpgrade) {
-      return acc;
-    }
-
-    const status = isSupported
-      ? AtomicCapabilityStatus.Supported
-      : AtomicCapabilityStatus.Ready;
-
-    if (acc[chainId as Hex] === undefined) {
-      acc[chainId as Hex] = {};
-    }
-
-    acc[chainId as Hex].atomic = {
-      status,
-    };
-
-    return acc;
-  }, alternateGasFeesAcc);
-}
-
-/**
  * Processes a single transaction from a sendCalls request.
  *
  * @param params - Object containing all parameters needed for single transaction processing.
@@ -366,7 +167,7 @@ async function processSingleTransaction({
   };
 
   const securityRequest: ValidateSecurityRequest = {
-    method: MESSAGE_TYPE.ETH_SEND_TRANSACTION,
+    method: MessageType.SendTransaction,
     params: [txParams],
     origin,
   };
@@ -633,130 +434,4 @@ function validateUpgrade(
       'EIP-7702 upgrade not supported on account',
     );
   }
-}
-
-/**
- * Determines alternate gas fees capability for the specified chains.
- *
- * @param chainIds - Array of chain IDs to check for alternate gas fees support.
- * @param batchSupport - Information about atomic batch support for each chain.
- * @param getIsSmartTransaction - Function to check if a chain supports smart transactions.
- * @param isRelaySupported - Function to check if relay is supported on a chain.
- * @param getSendBundleSupportedChains - Function to get chains that support send bundle.
- * @param messenger - Messenger instance for controller communication.
- * @returns Promise resolving to GetCapabilitiesResult with alternate gas fees information.
- */
-async function getAlternateGasFeesCapability(
-  chainIds: Hex[],
-  batchSupport: IsAtomicBatchSupportedResult,
-  getIsSmartTransaction: (chainId: Hex) => boolean,
-  isRelaySupported: (chainId: Hex) => Promise<boolean>,
-  getSendBundleSupportedChains: (
-    chainIds: Hex[],
-  ) => Promise<Record<string, boolean>>,
-  messenger: EIP5792Messenger,
-) {
-  const simulationEnabled = messenger.call(
-    'PreferencesController:getState',
-  ).useTransactionSimulations;
-
-  const relaySupportedChains = await Promise.all(
-    batchSupport
-      .map(({ chainId }) => chainId)
-      .map((chainId) => isRelaySupported(chainId)),
-  );
-
-  const sendBundleSupportedChains =
-    await getSendBundleSupportedChains(chainIds);
-
-  const updatedBatchSupport = batchSupport.map((support, index) => ({
-    ...support,
-    relaySupportedForChain: relaySupportedChains[index],
-  }));
-
-  return chainIds.reduce<GetCapabilitiesResult>((acc, chainId) => {
-    const chainBatchSupport = (updatedBatchSupport.find(
-      ({ chainId: batchChainId }) => batchChainId === chainId,
-    ) ?? {}) as IsAtomicBatchSupportedResultEntry & {
-      relaySupportedForChain: boolean;
-    };
-
-    const { isSupported = false, relaySupportedForChain } = chainBatchSupport;
-
-    const isSmartTransaction = getIsSmartTransaction(chainId);
-    const isSendBundleSupported = sendBundleSupportedChains[chainId] ?? false;
-
-    const alternateGasFees =
-      simulationEnabled &&
-      ((isSmartTransaction && isSendBundleSupported) ||
-        (isSupported && relaySupportedForChain));
-
-    if (alternateGasFees) {
-      acc[chainId as Hex] = {
-        alternateGasFees: {
-          supported: true,
-        },
-      };
-    }
-
-    return acc;
-  }, {});
-}
-
-/**
- * Maps transaction status to EIP-5792 call status codes.
- *
- * @param transactionMeta - The transaction metadata containing status and hash information.
- * @returns GetCallsStatusCode representing the current status of the transaction.
- */
-function getStatusCode(transactionMeta: TransactionMeta) {
-  const { hash, status } = transactionMeta;
-
-  if (status === TransactionStatus.confirmed) {
-    return GetCallsStatusCode.CONFIRMED;
-  }
-
-  if (status === TransactionStatus.failed) {
-    return hash
-      ? GetCallsStatusCode.REVERTED
-      : GetCallsStatusCode.FAILED_OFFCHAIN;
-  }
-
-  if (status === TransactionStatus.dropped) {
-    return GetCallsStatusCode.REVERTED;
-  }
-
-  return GetCallsStatusCode.PENDING;
-}
-
-/**
- * Retrieves the keyring type for a given account address.
- *
- * @param accountAddress - The account address to look up.
- * @param messenger - Messenger instance for controller communication.
- * @returns The keyring type associated with the account.
- * @throws JsonRpcError if the account type is unknown or not found.
- */
-function getAccountKeyringType(
-  accountAddress: Hex,
-  messenger: EIP5792Messenger,
-): KeyringTypes {
-  const { accounts } = messenger.call(
-    'AccountsController:getState',
-  ).internalAccounts;
-
-  const account = Object.values(accounts).find(
-    (acc) => acc.address.toLowerCase() === accountAddress.toLowerCase(),
-  );
-
-  const keyringType = account?.metadata?.keyring?.type;
-
-  if (!keyringType) {
-    throw new JsonRpcError(
-      EIP5792ErrorCode.RejectedUpgrade,
-      'EIP-7702 upgrade not supported as account type is unknown',
-    );
-  }
-
-  return keyringType as KeyringTypes;
 }
