+++ packages/eip-5792-middleware/src/types.ts
@@ -1,0 +1,21 @@
+import type {
+  AccountsControllerGetSelectedAccountAction,
+  AccountsControllerGetStateAction,
+} from '@metamask/accounts-controller';
+import type { Messenger } from '@metamask/base-controller';
+import type {
+  NetworkControllerGetNetworkClientByIdAction,
+  NetworkControllerGetStateAction,
+} from '@metamask/network-controller';
+import type { PreferencesControllerGetStateAction } from '@metamask/preferences-controller';
+import type { TransactionControllerGetStateAction } from '@metamask/transaction-controller';
+
+type Actions =
+  | AccountsControllerGetStateAction
+  | AccountsControllerGetSelectedAccountAction
+  | NetworkControllerGetNetworkClientByIdAction
+  | TransactionControllerGetStateAction
+  | PreferencesControllerGetStateAction
+  | NetworkControllerGetStateAction;
+
+export type EIP5792Messenger = Messenger<Actions, never>;
